bruin/requirements_bruin.txt
pandas>=2.2
psycopg2-binary>=2.9
openpyxl>=3.1
python-dotenv>=1.0
